+++ micrometer1-shim/src/main/java/io/opentelemetry/micrometershim/DoubleMeasurementRecorder.java
