+++ micrometer1-shim/src/main/java/io/opentelemetry/micrometershim/OpenTelemetryFunctionTimer.java
