+++ micrometer1-shim/src/main/java/io/opentelemetry/micrometershim/RemovableMeter.java
