+++ micrometer1-shim/src/test/java/io/opentelemetry/micrometershim/CounterTest.java
