+++ micrometer1-shim/src/test/java/io/opentelemetry/micrometershim/FunctionCounterTest.java
