+++ micrometer1-shim/src/test/java/io/opentelemetry/micrometershim/FunctionTimerSecondsTest.java
@@ -9,81 +9,84 @@ import static io.opentelemetry.micrometershim.OpenTelemetryMeterRegistryBuilder.
 import static io.opentelemetry.sdk.testing.assertj.MetricAssertions.assertThat;
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.attributeEntry;
 
+import io.micrometer.core.instrument.FunctionTimer;
 import io.micrometer.core.instrument.Metrics;
-import io.micrometer.core.instrument.Timer;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import java.util.concurrent.TimeUnit;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-@SuppressWarnings("PreferJavaTimeOverload")
-class TimerSecondsTest {
+class FunctionTimerSecondsTest {
 
   @RegisterExtension
   static final MicrometerTestingExtension testing =
       new MicrometerTestingExtension() {
 
         @Override
-        OpenTelemetryMeterRegistryBuilder configureRegistry(
+        OpenTelemetryMeterRegistryBuilder configureOtelRegistry(
             OpenTelemetryMeterRegistryBuilder registry) {
           return registry.setBaseTimeUnit(TimeUnit.SECONDS);
         }
       };
 
+  final TestTimer timerObj = new TestTimer();
+
+  @BeforeEach
+  void cleanupTimer() {
+    timerObj.reset();
+  }
+
   @Test
-  void testTimerWithBaseUnitSeconds() {
-    Timer timer =
-        Timer.builder("testTimerSeconds")
-            .description("This is a test timer")
+  void testFunctionTimerWithBaseUnitSeconds() {
+    FunctionTimer functionTimer =
+        FunctionTimer.builder(
+                "testFunctionTimerSeconds",
+                timerObj,
+                TestTimer::getCount,
+                TestTimer::getTotalTimeNanos,
+                TimeUnit.NANOSECONDS)
+            .description("This is a test function timer")
             .tags("tag", "value")
             .register(Metrics.globalRegistry);
 
-    timer.record(1, TimeUnit.SECONDS);
-    timer.record(10, TimeUnit.SECONDS);
-    timer.record(12_345, TimeUnit.MILLISECONDS);
+    timerObj.add(42, TimeUnit.SECONDS);
 
     assertThat(testing.collectAllMetrics())
         .satisfiesExactlyInAnyOrder(
             metric ->
                 assertThat(metric)
-                    .hasName("testTimerSeconds")
+                    .hasName("testFunctionTimerSeconds.count")
                     .hasInstrumentationLibrary(
                         InstrumentationLibraryInfo.create(INSTRUMENTATION_NAME, null))
-                    .hasDescription("This is a test timer")
-                    .hasUnit("s")
-                    .hasDoubleHistogram()
+                    .hasDescription("This is a test function timer")
+                    .hasUnit("1")
+                    .hasLongSum()
+                    .isMonotonic()
                     .points()
                     .satisfiesExactly(
                         point ->
                             assertThat(point)
-                                .hasSum(23.345)
-                                .hasCount(3)
+                                .hasValue(1)
                                 .attributes()
                                 .containsOnly(attributeEntry("tag", "value"))),
             metric ->
                 assertThat(metric)
-                    .hasName("testTimerSeconds.max")
-                    .hasDescription("This is a test timer")
+                    .hasName("testFunctionTimerSeconds.sum")
+                    .hasInstrumentationLibrary(
+                        InstrumentationLibraryInfo.create(INSTRUMENTATION_NAME, null))
+                    .hasDescription("This is a test function timer")
                     .hasUnit("s")
-                    .hasDoubleGauge()
+                    .hasDoubleSum()
                     .points()
-                    .anySatisfy(
+                    .satisfiesExactly(
                         point ->
                             assertThat(point)
-                                .hasValue(12.345)
+                                .hasValue(42)
                                 .attributes()
-                                .containsEntry("tag", "value")));
+                                .containsOnly(attributeEntry("tag", "value"))));
 
-    Metrics.globalRegistry.remove(timer);
-    timer.record(12, TimeUnit.SECONDS);
-
-    assertThat(testing.collectAllMetrics())
-        .satisfiesExactlyInAnyOrder(
-            metric ->
-                assertThat(metric)
-                    .hasName("testTimerSeconds")
-                    .hasDoubleHistogram()
-                    .points()
-                    .noneSatisfy(point -> assertThat(point).hasSum(35.345).hasCount(4)));
+    Metrics.globalRegistry.remove(functionTimer);
+    assertThat(testing.collectAllMetrics()).isEmpty();
   }
 }
