+++ settings.gradle.kts
@@ -53,7 +53,7 @@ include(":exporters:zipkin")
 include(":integration-tests")
 include(":integration-tests:otlp")
 include(":integration-tests:tracecontext")
-include(":micrometer-shim")
+include(":micrometer1-shim")
 include(":opencensus-shim")
 include(":opentracing-shim")
 include(":perf-harness")
